backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai>=1.51.0
python-dotenv>=1.0
supabase>=2.4.0
psycopg[binary]>=3.2
PyJWT>=2.10
